Implement search feature Get and display Repositories Data Configure RepositoriesViewController Configure RepoTableViewCell Add Repositories. model and update repository model Update config file Update GithubbRouter with parameters
