+++ GithubClient/Models/Repository.swift
@@ -1,0 +1,127 @@
+//
+//  Repository.swift
+//  GithubClient
+//
+//  Created by Farghaly on 16/11/2021.
+//
+
+import Foundation
+
+// MARK: - Repository
+struct Repository: Codable {
+    let id: Int?
+    let nodeID, name, fullName: String?
+    let welcomePrivate: Bool?
+    let owner: Owner?
+    let htmlURL: String?
+    let welcomeDescription: String?
+    let fork: Bool?
+    let url, forksURL: String?
+    let keysURL, collaboratorsURL: String?
+    let teamsURL, hooksURL: String?
+    let issueEventsURL: String?
+    let eventsURL: String?
+    let assigneesURL, branchesURL: String?
+    let tagsURL: String?
+    let blobsURL, gitTagsURL, gitRefsURL, treesURL: String?
+    let statusesURL: String?
+    let languagesURL, stargazersURL, contributorsURL, subscribersURL: String?
+    let subscriptionURL: String?
+    let commitsURL, gitCommitsURL, commentsURL, issueCommentURL: String?
+    let contentsURL, compareURL: String?
+    let mergesURL: String?
+    let archiveURL: String?
+    let downloadsURL: String?
+    let issuesURL, pullsURL, milestonesURL, notificationsURL: String?
+    let labelsURL, releasesURL: String?
+    let deploymentsURL: String?
+
+    enum CodingKeys: String, CodingKey {
+        case id
+        case nodeID = "node_id"
+        case name
+        case fullName = "full_name"
+        case welcomePrivate = "private"
+        case owner
+        case htmlURL = "html_url"
+        case welcomeDescription = "description"
+        case fork, url
+        case forksURL = "forks_url"
+        case keysURL = "keys_url"
+        case collaboratorsURL = "collaborators_url"
+        case teamsURL = "teams_url"
+        case hooksURL = "hooks_url"
+        case issueEventsURL = "issue_events_url"
+        case eventsURL = "events_url"
+        case assigneesURL = "assignees_url"
+        case branchesURL = "branches_url"
+        case tagsURL = "tags_url"
+        case blobsURL = "blobs_url"
+        case gitTagsURL = "git_tags_url"
+        case gitRefsURL = "git_refs_url"
+        case treesURL = "trees_url"
+        case statusesURL = "statuses_url"
+        case languagesURL = "languages_url"
+        case stargazersURL = "stargazers_url"
+        case contributorsURL = "contributors_url"
+        case subscribersURL = "subscribers_url"
+        case subscriptionURL = "subscription_url"
+        case commitsURL = "commits_url"
+        case gitCommitsURL = "git_commits_url"
+        case commentsURL = "comments_url"
+        case issueCommentURL = "issue_comment_url"
+        case contentsURL = "contents_url"
+        case compareURL = "compare_url"
+        case mergesURL = "merges_url"
+        case archiveURL = "archive_url"
+        case downloadsURL = "downloads_url"
+        case issuesURL = "issues_url"
+        case pullsURL = "pulls_url"
+        case milestonesURL = "milestones_url"
+        case notificationsURL = "notifications_url"
+        case labelsURL = "labels_url"
+        case releasesURL = "releases_url"
+        case deploymentsURL = "deployments_url"
+    }
+}
+
+// MARK: - Owner
+struct Owner: Codable {
+    let login: String?
+    let id: Int?
+    let nodeID: String?
+    let avatarURL: String?
+    let gravatarID: String?
+    let url, htmlURL, followersURL: String?
+    let followingURL, gistsURL, starredURL: String?
+    let subscriptionsURL, organizationsURL, reposURL: String?
+    let eventsURL: String?
+    let receivedEventsURL: String?
+    let type: TypeEnum?
+    let siteAdmin: Bool?
+
+    enum CodingKeys: String, CodingKey {
+        case login, id
+        case nodeID = "node_id"
+        case avatarURL = "avatar_url"
+        case gravatarID = "gravatar_id"
+        case url
+        case htmlURL = "html_url"
+        case followersURL = "followers_url"
+        case followingURL = "following_url"
+        case gistsURL = "gists_url"
+        case starredURL = "starred_url"
+        case subscriptionsURL = "subscriptions_url"
+        case organizationsURL = "organizations_url"
+        case reposURL = "repos_url"
+        case eventsURL = "events_url"
+        case receivedEventsURL = "received_events_url"
+        case type
+        case siteAdmin = "site_admin"
+    }
+}
+
+enum TypeEnum: String, Codable {
+    case organization = "Organization"
+    case user = "User"
+}
