+++ GithubClient/Scenes/Repositories/Service/RepositoriesService.swift
@@ -9,12 +9,18 @@ import Foundation
 import RxSwift
 
 protocol RepositoriesService {
-    func getRepositories() -> Observable<Repositories>
+    func searchRepositories(key: String) -> Observable<Repositories>
+    func getRepositories() -> Observable<[Repository]>
 }
 
 class RepositoriesServiceImpl: RepositoriesService {
     private let service = NetworkService()
-    func getRepositories() -> Observable<Repositories> {
+
+    func searchRepositories(key: String) -> Observable<Repositories> {
+        return service.fetchRequest(forRoute: .search(key: key))
+    }
+
+    func getRepositories() -> Observable<[Repository]> {
         return service.fetchRequest(forRoute: .getRepos)
     }
 }
