+++ GithubClient/Scenes/Repositories/View/XIBS/RepoTableViewCell.swift
@@ -11,15 +11,24 @@ class RepoTableViewCell: UITableViewCell {
 
     @IBOutlet weak var repoOwnerImage: UIImageView!
     @IBOutlet weak var repoName: UILabel!
+    @IBOutlet weak var repoOwnerName: UILabel!
     @IBOutlet weak var creationDate: UILabel!
+
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
     }
 
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
+    var viewModel: RepoViewModel! {
+        didSet {
+            self.setup()
+        }
+    }
 
-        // Configure the view for the selected state
+}
+
+private extension RepoTableViewCell {
+    func setup() {
+        repoName.text = viewModel.name
     }
 }
