+++ Podfile
@@ -14,3 +14,11 @@ target 'GithubClient' do
     # Pods for testing
   end
 end
+
+post_install do |installer|
+ installer.pods_project.targets.each do |target|
+  target.build_configurations.each do |config|
+   config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '13.0'
+  end
+ end
+end
